&&一些警告的处理

+++ CrossApp/control/CABar.cpp
@@ -348,7 +348,7 @@ bool CATabBar::init(const CAVector<CATabBarItem*>& items, const CCSize& size)
     m_pContentView->setCenter(rect);
     this->addSubview(m_pContentView);
     
-    unsigned int count = m_pItems.size();
+    unsigned int count = (unsigned int)m_pItems.size();
     m_cItemSize = m_pContentView->getBounds().size;
     m_cItemSize.width = rect.size.width/count;
