+++ CrossApp/view/CATableView.cpp
@@ -568,9 +568,9 @@ void CATableView::loadTableCell()
     rect.origin.y -= rect.size.height * 0.1f;
     rect.size.height *= 1.2f;
     
-    for (size_t i=0; i<m_rTableCellRectss.size(); i++)
+    for (unsigned int i=0; i<(unsigned int)m_rTableCellRectss.size(); i++)
     {
-        for (size_t j=0; j<m_rTableCellRectss.at(i).size(); j++)
+        for (unsigned int j=0; j<(unsigned int)m_rTableCellRectss.at(i).size(); j++)
         {
             CAIndexPath2E indexPath = CAIndexPath2E(i, j);
             CC_CONTINUE_IF(m_pUsedTableCells.count(indexPath) && m_pUsedTableCells[indexPath]);
@@ -653,7 +653,7 @@ void CATableView::updateSectionHeaderAndFooterRects()
     for (itr=m_rSectionRects.begin(); itr!=m_rSectionRects.end(); itr++)
     {
         CC_CONTINUE_IF(!rect.intersectsRect(*itr));
-        int i = itr - m_rSectionRects.begin();
+        int i = (int)(itr - m_rSectionRects.begin());
         CAView* header = NULL;
         CAView* footer = NULL;
         float headerHeight = m_nSectionHeaderHeights[i];
